Allow recovery if the transfer size is too large
I don't believe the camera ever sends a short or zero-length packet for
framing purposes, so we have to frame it ourselves, and that code is
somewhat buggy if any of the constants are defined incorrectly.
TRANSFER_SIZE > FRAME_PADDED_SIZE is one such case, it may write past
the end of our buffer.