Handle the possibility of pixel 0 being dead
Copying the value of the pixel before it won't work.  And in general
this won't work well for pixels at the left or right edge, so instead
set them to black.  If nearby pixels are dark then it won't be
noticeable; if they're light then their intensity may make the black
pixel hard to spot.