Move transfer allocation outside of the thread
Previously all transfers were either immediately resubmitted or
deallocated.  This will help create a third "idle" state where a
transfer is ready and waiting to be submitted.

The main problem is determining whether the USB thread has work to do
(i.e. whether libusb_handle_events() has something worth waiting for).
Ideally there would be a libusb function call to figure out if any
transfers are in-flight or completed/cancelled but not reaped.  Since
that doesn't seem to exist, we use the transfer's buffer pointer as a
proxy.  If any are non-null, the thread will eventually have work to do.