Update support for odd frame width/height
V4L2 docs say the Y plane width/height is padded to a multiple of 2, but
ffmpeg won't play it when it is.  Instead it expects no padding.  I
don't know who is right, but we weren't accounting for any pad rows/cols
in the rendering loop.  Do the simple thing and remove the padding so
that it plays in ffmpeg.  Also rounding the width up to 8 for the U/V
planes looks wrong, which might have been a botched attempt to find the
width of both planes.

Add support for NV12, NV21, and GREY which all should be compatible with
the current rendering loop.  Add unsupported conversions for VYUY and
Y14.  Drop Y41P whose size was calculated too small.