Header fields 0xb/0xc are the image height/width
When set to 0x0:  Output appears to be a 512x308 image containing four
smaller, horizontally-stretched, misaligned frames.  The bottom frame is
chronologically first.

When either dimension is set to < 80:  Output is the full 384x288 image.

When set to < 384x288:  Output is cropped from the center of the 384x288
image.  If either dimension is odd, one fewer row/col is removed from
the bottom/left side of the full image than the top/right (pre-flip).

When either dimension is set to > the corresponding sensor dimension:
The value returned in the header is clipped to the sensor dimension.

Changing image dimensions on-the-fly is not well supported.  Increasing
the height tends to leave a lighter or darker horizontal stripe of the
previous height across the center of the image.