Rename and update USB size constants
"chunk" is now "packet", derived from wMaxPacketSize in the descriptor
and related USB terminology.

TRANSFER_SIZE and FRAME_PADDED_SIZE are now BULK_SIZE_MIN and
BULK_SIZE_MAX, respectively.  These are the smallest and largest
supported frame/image sizes, where "frame" refers to link-layer data
framing, not an image frame, but in this case the terms are identical.
The minimum represents an 80x80 image and should be the largest transfer
that contains no more than one start-of-frame preamble.

Note that BULK_SIZE_MIN increases from 16 to 26 packets which should
reduce the number of wakeups while attempting to sync.  Further reduce
the number of wakeups once sync'd by requesting the exact number of
packets to complete a frame (where the expected frame size is currently
BULK_SIZE_MAX).