Add code to calculate header field 0x12
This doesn't yet send the updated value to the camera.  I may change it
to return a delta, in case we don't need to send it if the delta is 0.
Note that the calculation is done on the raw data, including bad pixels.

Version format 0:  min/max/target are not present in the calibration,
always use the hard-coded defaults of 1392/2949/0.5 to calculate the new
value.

Version format 1:  Load the min/max/target values from the selected
calibration set.  To be calculated on each frame as long as target != 0
(which I am also using to skip invalid calibration sets).

Version format 2:  Same as 1, but since this indicates a 640x480 camera,
the calculation is skipped (even though the code has some decisions to
support this size, including ignoring the min/max calibration values).