Add histogram equalization and initial palette support
Static offset of 5000 is taken from the app.  I assume the units are
centi-centigrade which means the domain for the LUT is -50.00 C to
+605.35 C.

Fixed-point scaling is also taken from the app, and it loses precision
with large n and fails entirely when n > 0xff00.  That situation is
unlikely but I may rewrite it later.