Allow selecting enhanced mode from the command line
Support for the optional argument is a GNU extension.

getopt doesn't allow a space between the "-e" and the argument,
because to support it properly it would have to determine if the next
argv-element is the argument, or the start of another option, or a
non-option.

Enhanced mode tends to not work well with the automatic calibration
since the image has a narrow dynamic range.